TeachingAgent/backend/pyproject.toml
[project]
name = "teaching-agent-backend"
version = "0.1.0"
description = "Local-first adaptive learning engine backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.35",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.2",
  "python-multipart>=0.0.9",
  "httpx>=0.27.2",
  "networkx>=3.3",
  "numpy>=2.1.1",
  "orjson>=3.10.7",
  "python-docx>=1.1.2",
  "pypdf>=4.3.1",
  "lxml>=5.3.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["app"]
